+++ src/FluentMigrator.Runner.Postgres/PostgresRunnerBuilderExtensions.cs
@@ -1,4 +1,4 @@
-﻿#region License
+#region License
 // Copyright (c) 2018, FluentMigrator Project
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -66,13 +66,13 @@ namespace FluentMigrator.Runner
         /// </summary>
         /// <param name="builder">The builder to add the Postgres-specific services to</param>
         /// <returns>The migration runner builder</returns>
-        public static IMigrationRunnerBuilder AddPostgres100(this IMigrationRunnerBuilder builder)
+        public static IMigrationRunnerBuilder AddPostgres10_0(this IMigrationRunnerBuilder builder)
         {
             builder.Services
-                .AddScoped<Postgres100Processor>()
-                .AddScoped<IMigrationProcessor>(sp => sp.GetRequiredService<Postgres100Processor>())
-                .AddScoped<Postgres100Generator>()
-                .AddScoped<IMigrationGenerator>(sp => sp.GetRequiredService<Postgres100Generator>());
+                .AddScoped<Postgres10_0Processor>()
+                .AddScoped<IMigrationProcessor>(sp => sp.GetRequiredService<Postgres10_0Processor>())
+                .AddScoped<Postgres10_0Generator>()
+                .AddScoped<IMigrationGenerator>(sp => sp.GetRequiredService<Postgres10_0Generator>());
             return builder.AddCommonPostgresServices();
         }
